Add files via upload
I have been having some issues regarding the long term impact of the wage mark-up shock. These issues concerning the impact on the long term variance decomposition of inflation and output might be explained due to the bias that the financial crisis creates on my short sample of years (1999-2019). The code provided takes the subset 1999-2006 of the series, fit an ARIMA model for each of the series and then forecast the series until 2019. The purpose of doing this is to construct a database without the impact of the financial crisis an then run again the estimation to see if there are significant changes in the variance decomposition of the variables.